commit2

+++ src/main/java/com/example/bot/Telegram_bot_take_it/admin/dto/DashboardStatsDto.java
@@ -31,5 +31,5 @@ public class DashboardStatsDto {
     private Integer ordersToday;
 
     /** Выручка за сегодня */
-    private Integer todayRevenue;
+    private Long todayRevenue;
 }
